# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/exercise_group": {
            "get": {
                "tags": [
                    "exercise group"
                ],
                "summary": "get group / get group for a specific date.",
                "description": "get group / get group for a specific date.",
                "operationId": "12cc1db64a84dd64f45b65090ae6b34e",
                "parameters": [
                    {
                        "name": "date",
                        "in": "query",
                        "description": "search by month(01,02,03,04....ex) example:2021-10",
                        "required": false
                    },
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "search by user_id",
                        "required": false
                    },
                    {
                        "name": "plan_id",
                        "in": "query",
                        "description": "search by plan_id",
                        "required": false
                    },
                    {
                        "name": "trainer_id",
                        "in": "query",
                        "description": "search by trainer_id",
                        "required": false
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "search by page",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "exercise group"
                ],
                "summary": "create new exercise group",
                "description": "create new exercise group",
                "operationId": "519a5e61999483dbe4dfab54125e4c01",
                "requestBody": {
                    "description": "create exercise group",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "date": {
                                        "type": "string",
                                        "example": "2021-12-4"
                                    },
                                    "plan_id": {
                                        "description": "plan's id",
                                        "type": "number"
                                    },
                                    "exercises": {
                                        "type": "string",
                                        "example": "[{'name':'test','sets':'20 20'}]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/exercise_group/{group_id}": {
            "put": {
                "tags": [
                    "exercise group"
                ],
                "summary": "update exercise group",
                "description": "update exercise group",
                "operationId": "5d713e0c800da66351857eab8461cb9b",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "description": "group_id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "update exercise group",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "description": {
                                        "type": "string",
                                        "example": "test"
                                    },
                                    "date": {
                                        "type": "string",
                                        "example": "2021-12-4"
                                    },
                                    "plan_id": {
                                        "description": "plan's id",
                                        "type": "number"
                                    },
                                    "exercises": {
                                        "type": "string",
                                        "example": "[{'name':'test','sets':'20 20'}]"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "exercise group"
                ],
                "summary": "delete exercise group by id",
                "description": "delete exercise group by id",
                "operationId": "6d61ac461d93c99189acbef35aac1328",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "path",
                        "description": "delete group exercise by id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}